+++ src/api/android/mod.rs
@@ -99,9 +99,9 @@ impl Context {
         _el: &winit::EventsLoop,
         pf_reqs: &PixelFormatRequirements,
         gl_attr: &GlAttributes<&Context>,
-        shareable_with_windowed_contextes: bool,
+        shareable_with_windowed_contexts: bool,
     ) -> Result<Self, CreationError> {
-        assert!(shareable_with_windowed_contextes); // TODO: Implement if possible
+        assert!(shareable_with_windowed_contexts); // TODO: Implement if possible
 
         let gl_attr = gl_attr.clone().map_sharing(|c| &c.0);
         let context = EglContext::new(